+++ R/multi.R
@@ -5,7 +5,7 @@ multiX <- function(X, m) {
   for (i in 1:m) {
     A[m*(1:n)-i+1, m*(1:p)-i+1] <- X
   }
-  cbind(matrix(as.numeric(diag(m)),m*n,m,byrow=TRUE)[,2:m],A)
+  cbind(matrix(as.double(diag(m)), m*n, m, byrow=TRUE)[,2:m], A)
 }
 multiG <- function(g, ncolY) {
   structure(c(rep(0, ncolY-1), rep(g, each=ncolY)),
